Move .buildings.Config to a dedicated submodule
Mirror structure in .model.{material,transport}.
